+++ app/src/main/java/org/gateshipone/odyssey/activities/OdysseyMainActivity.java
@@ -64,6 +64,7 @@ import android.widget.TextView;
 
 import org.gateshipone.odyssey.R;
 import org.gateshipone.odyssey.adapter.CurrentPlaylistAdapter;
+import org.gateshipone.odyssey.dialogs.SaveDialog;
 import org.gateshipone.odyssey.fragments.AlbumTracksFragment;
 import org.gateshipone.odyssey.fragments.ArtistAlbumsFragment;
 import org.gateshipone.odyssey.fragments.ArtworkSettingsFragment;
@@ -72,7 +73,7 @@ import org.gateshipone.odyssey.fragments.FilesFragment;
 import org.gateshipone.odyssey.fragments.MyMusicFragment;
 import org.gateshipone.odyssey.fragments.OdysseyFragment;
 import org.gateshipone.odyssey.fragments.PlaylistTracksFragment;
-import org.gateshipone.odyssey.dialogs.SaveDialog;
+import org.gateshipone.odyssey.fragments.RecentAlbumsFragment;
 import org.gateshipone.odyssey.fragments.SavedPlaylistsFragment;
 import org.gateshipone.odyssey.fragments.SettingsFragment;
 import org.gateshipone.odyssey.listener.OnAlbumSelectedListener;
@@ -80,6 +81,7 @@ import org.gateshipone.odyssey.listener.OnArtistSelectedListener;
 import org.gateshipone.odyssey.listener.OnDirectorySelectedListener;
 import org.gateshipone.odyssey.listener.OnPlaylistFileSelectedListener;
 import org.gateshipone.odyssey.listener.OnPlaylistSelectedListener;
+import org.gateshipone.odyssey.listener.OnRecentAlbumsSelectedListener;
 import org.gateshipone.odyssey.listener.OnSaveDialogListener;
 import org.gateshipone.odyssey.listener.ToolbarAndFABCallback;
 import org.gateshipone.odyssey.models.AlbumModel;
@@ -96,8 +98,10 @@ import org.gateshipone.odyssey.views.NowPlayingView;
 import java.util.List;
 
 public class OdysseyMainActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener, OnArtistSelectedListener, OnAlbumSelectedListener, OnPlaylistSelectedListener, OnSaveDialogListener,
-        OnDirectorySelectedListener, OnPlaylistFileSelectedListener, NowPlayingView.NowPlayingDragStatusReceiver, SettingsFragment.OnArtworkSettingsRequestedCallback, ToolbarAndFABCallback {
+        implements NavigationView.OnNavigationItemSelectedListener, ToolbarAndFABCallback,
+        OnSaveDialogListener, NowPlayingView.NowPlayingDragStatusReceiver, SettingsFragment.OnArtworkSettingsRequestedCallback,
+        OnArtistSelectedListener, OnAlbumSelectedListener, OnRecentAlbumsSelectedListener,
+        OnPlaylistSelectedListener, OnPlaylistFileSelectedListener, OnDirectorySelectedListener {
 
     private ActionBarDrawerToggle mDrawerToggle;
 
@@ -483,6 +487,20 @@ public class OdysseyMainActivity extends AppCompatActivity
             }
 
             CurrentPlaylistView currentPlaylistView = (CurrentPlaylistView) findViewById(R.id.now_playing_playlist);
+
+            // check if track has a valid album key
+            String albumKey = currentPlaylistView.getAlbumKey(info.position);
+            AlbumModel tmpAlbum = MusicLibraryHelper.createAlbumModelFromKey(albumKey, getApplicationContext());
+
+            menu.findItem(R.id.view_current_playlist_action_showalbum).setVisible(tmpAlbum != null);
+
+            // check if track has a valid artist id
+            String artistTitle = currentPlaylistView.getArtistTitle(info.position);
+            long artistID = MusicLibraryHelper.getArtistIDFromName(artistTitle, this);
+
+            menu.findItem(R.id.view_current_playlist_action_showartist).setVisible(artistID != -1);
+
+            // check the view type
             if (currentPlaylistView.getItemViewType(info.position) == CurrentPlaylistAdapter.VIEW_TYPES.TYPE_SECTION_TRACK_ITEM) {
                 menu.findItem(R.id.view_current_playlist_action_remove_section).setVisible(true);
             }
@@ -945,6 +963,28 @@ public class OdysseyMainActivity extends AppCompatActivity
         if (collapsingImage != null) {
             collapsingImage.setImageBitmap(bm);
         }
+    }
+
+    @Override
+    public void onRecentAlbumsSelected() {
+        // Create fragment
+        RecentAlbumsFragment newFragment = new RecentAlbumsFragment();
+
+        android.support.v4.app.FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+
+        // set enter / exit animation
+        newFragment.setEnterTransition(new Slide(Gravity.BOTTOM));
+        newFragment.setExitTransition(new Slide(Gravity.TOP));
+
+        // Replace whatever is in the fragment_container view with this
+        // fragment,
+        // and add the transaction to the back stack so the user can navigate
+        // back
+        transaction.replace(R.id.fragment_container, newFragment);
+        transaction.addToBackStack("RecentAlbumsFragment");
+
+        // Commit the transaction
+        transaction.commit();
     }
 
     private MyMusicFragment.DEFAULTTAB getDefaultTab() {