+++ app/src/main/java/org/gateshipone/odyssey/fragments/AlbumsFragment.java
@@ -77,14 +77,15 @@ public class AlbumsFragment extends GenericAlbumsFragment {
 
     /**
      * This method creates a new loader for this fragment.
-     * @param id The id of the loader
+     *
+     * @param id     The id of the loader
      * @param bundle Optional arguments
      * @return Return a new Loader instance that is ready to start loading.
      */
     @Override
     public Loader<List<AlbumModel>> onCreateLoader(int id, Bundle bundle) {
         // all albums
-        return new AlbumLoader(getActivity());
+        return new AlbumLoader(getActivity(), false);
     }
 
     /**
@@ -99,6 +100,7 @@ public class AlbumsFragment extends GenericAlbumsFragment {
 
     /**
      * Hook called when an menu item in the context menu is selected.
+     *
      * @param item The menu item that was selected.
      * @return True if the hook was consumed here.
      */
@@ -127,6 +129,7 @@ public class AlbumsFragment extends GenericAlbumsFragment {
 
     /**
      * Open a fragment for the artist of the selected album.
+     *
      * @param position the position of the selected album in the adapter
      */
     private void showArtist(int position) {