+++ app/src/main/java/org/gateshipone/odyssey/loaders/PlaylistLoader.java
@@ -24,13 +24,10 @@ package org.gateshipone.odyssey.loaders;
 
 import android.support.v4.content.AsyncTaskLoader;
 import android.content.Context;
-import android.database.Cursor;
-import android.provider.MediaStore;
 
 import org.gateshipone.odyssey.R;
 import org.gateshipone.odyssey.models.PlaylistModel;
 import org.gateshipone.odyssey.utils.MusicLibraryHelper;
-import org.gateshipone.odyssey.utils.PermissionHelper;
 
 import java.util.ArrayList;
 import java.util.List;