+++ app/src/main/java/org/gateshipone/odyssey/playbackservice/OdysseyPlaybackServiceInterface.java
@@ -332,6 +332,14 @@ public class OdysseyPlaybackServiceInterface extends IOdysseyPlaybackService.Stu
     }
 
     @Override
+    public void enqueueRecentAlbums() throws RemoteException {
+        ControlObject obj = new ControlObject(ControlObject.PLAYBACK_ACTION.ODYSSEY_ENQUEUERECENTALBUMS);
+        Message msg = mService.get().getHandler().obtainMessage();
+        msg.obj = obj;
+        mService.get().getHandler().sendMessage(msg);
+    }
+
+    @Override
     public void enqueueArtist(long artistId, String orderKey) throws RemoteException {
         ControlObject obj = new ControlObject(ControlObject.PLAYBACK_ACTION.ODYSSEY_ENQUEUEARTIST, artistId, orderKey);
         Message msg = mService.get().getHandler().obtainMessage();